etl/requirements.txt
pandas == 1.1.4
nltk == 3.5
scikit-learn == 0.23.2
mongoengine== 0.20.0
celery==5.0.2
gevent==20.9.0
pytest==6.1.2
PyYAML==5.3.1
